feature: Adding meta properties to Section
Currently, you only get access to the `id` passed into `<Section>`s
when you retrieve section information later with the
`useScrollSections()` hook.  This allows you to pass through extra
information that might be useful.

In my case, I wanted to have a title that wasn't dependent on the ID
when using @EmaSuriano's
[gatsby-starter-mate](https://github.com/EmaSuriano/gatsby-starter-mate),
specifically so I could have titles with spaces in them and continue to
use internal markdown links.